+++ frontend/src/components/mod.rs
@@ -1,4 +1,5 @@
 pub mod about_modal;
+pub mod download_dat_modal;
 pub mod import_dat_modal;
 pub mod navbar;
 pub mod notifications;