+++ frontend/src/state.rs
@@ -106,6 +106,7 @@ pub struct AppState {
     pub name_filter: RwSignal<String>,
 
     // Modals.
+    pub download_dat_modal_open: RwSignal<bool>,
     pub import_dat_modal_open: RwSignal<bool>,
     pub settings_modal_open: RwSignal<bool>,
     pub about_modal_open: RwSignal<bool>,
@@ -316,6 +317,7 @@ impl AppState {
             one_region_filter,
             name_filter,
 
+            download_dat_modal_open: RwSignal::new(false),
             import_dat_modal_open: RwSignal::new(false),
             settings_modal_open: RwSignal::new(false),
             about_modal_open: RwSignal::new(false),